등록페이지와 수정페이지 분리
