+++ src/components/page-layout/helpMeEditLayout/components/helpMeEdit.scss
@@ -1,0 +1,206 @@
+.container {
+    margin: 9.8rem 0;
+    @include flexSort(null, center, null, null);
+  }
+  
+  .box {
+    @include flexSort(column, null, center, 5rem);
+    border: 0.1rem solid black;
+    width: 100rem;
+    border-radius: 1rem;
+    padding: 5rem 0;
+  }
+  
+  .title {
+    @include nsnFont(2.3rem, 700, normal);
+  }
+  
+  .form {
+    @include flexSort(column, null, center, 2rem);
+  }
+  
+  .formContentBox {
+    @include flexSort(column, null, center, 4rem);
+    width: 68rem;
+  }
+  
+  .titleContainer,
+  .dateContainer,
+  .timeContainer,
+  .periodContainer,
+  .placeHelpTypeContainer,
+  .detailContainer {
+    width: 100%;
+  }
+  
+  .label {
+    @include nsnFont(1.8rem, 800, normal);
+    padding: 0 1rem;
+  }
+  
+  .titleInput {
+    border-radius: 1rem;
+    @include nsnFont(1.8rem, 400, normal);
+    padding: 1rem 2rem;
+    width: calc(100% - 2rem);
+    border: 0.1rem solid #9e9696;
+    margin: 0 1rem;
+  
+    &::placeholder {
+      color: #b7b7b7;
+    }
+  }
+  
+  .errorMessage {
+    @include nsnFont(1.3rem, 500, normal);
+    color: red;
+    margin: 0.5rem 1rem 0 1rem;
+  }
+  
+  .dateBox {
+    width: 100%;
+    padding: 0 1rem;
+    @include flexSort(null, space-between, center, null);
+  }
+  
+  .date {
+    position: relative;
+    width: 31rem;
+    height: 4.48rem;
+  }
+  
+  .dateFont {
+    @include nsnFont(1.8rem, 400, normal);
+    &::placeholder {
+      color: #b7b7b7;
+    }
+  }
+  
+  .calendar {
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+  }
+  
+  .wave {
+    @include nsnFont(1.8rem, 800, normal);
+  }
+  
+  .timeBox {
+    @include flexSort(null, space-between, center, null);
+    padding: 0 1rem;
+  }
+  
+  .assistanceStartTimeBox,
+  .assistanceEndTimeBox,
+  .scheduleTypeBox,
+  .scheduleDetailsBox {
+    @include flexSort(column, null, null, null);
+  }
+  
+  .assistanceStartTime,
+  .assistanceEndTime {
+    width: 31rem;
+    height: 4.48rem;
+    border-radius: 1rem;
+    border: 0.1rem solid $gray;
+    @include nsnFont(1.8rem, 400, normal);
+    @include flexSort(null, center, center, null);
+    padding: 0 2rem;
+  }
+  
+  .periodBox {
+    @include flexSort(null, space-between, null, null);
+    padding: 0 1rem;
+  }
+  
+  .period {
+    height: 4.48rem;
+    width: 31rem;
+  }
+  
+  .periodDetailInput {
+    border-radius: 1rem;
+    border: 0.1rem solid $gray;
+    padding: 1rem 2rem;
+    @include nsnFont(1.8rem, 400, normal);
+    width: 31rem;
+    height: 4.48rem;
+  
+    &::placeholder {
+      color: #b7b7b7;
+    }
+  }
+  
+  .placeHelpTypeBox {
+    padding: 0 1rem;
+    @include flexSort(null, space-between, center, null);
+  }
+  
+  .placeBox,
+  .helpTypeBox {
+    width: 31rem;
+    height: 4.48rem;
+  }
+  
+  .detailTextarea {
+    resize: none;
+    width: 100%;
+    height: 22rem;
+    @include nsnFont(1.8rem, 400, normal);
+    overflow-y: auto;
+    padding: 2rem;
+    border-radius: 1rem;
+  
+    &::placeholder {
+      color: #b7b7b7;
+    }
+  }
+  
+  .registerBox {
+    padding: 0.7rem 1.4rem;
+    border-radius: 1rem;
+    @include nsnFont(2rem, 400, normal);
+    @include flexSort(null, null, center, 1rem);
+    color: #9d9d9d;
+    border: 0.1rem solid #9d9d9d;
+  
+    &.active {
+      color: #6562e2;
+      border: 0.1rem solid #6562e2;
+  
+      .arrow {
+        fill: #6562e2;
+      }
+    }
+  }
+  
+  .modal {
+    width: 75rem;
+    height: 21.8rem;
+    @include flexSort(column, center, center, 2.4rem);
+  }
+  
+  .textBox {
+    @include flexSort(column, null, center, 1rem);
+  }
+  
+  .modalTitle {
+    @include nsnFont(2rem, 700, normal);
+    @include flexSort(column, null, center, null);
+  }
+  
+  .modalContent {
+    @include nsnFont(1.4rem, 700, normal);
+    color: #5c5b5b;
+  }
+  
+  .modalBtnContent {
+    @include font(1.5rem, 700, normal);
+    padding: 0.75rem 4.55rem;
+    border-radius: 0.5rem;
+    border: 0.1rem solid #e6e6ea;
+    background: #e6e6ea;
+    box-shadow: 0 0.4rem 0.4rem 0 rgba(0, 0, 0, 0.25);
+  }
+
