+++ src/components/page-layout/helpMeRegisterLayout/components/helpMeRegisterLayout.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
 
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQuery } from "@tanstack/react-query";
 import classNames from "classnames/bind";
 import { ko } from "date-fns/locale";
 import { Controller, useForm } from "react-hook-form";
@@ -25,10 +25,8 @@ import { ROUTE } from "@/constants/route";
 import Calendar from "@/icons/calendar.svg";
 import RegisterArrow from "@/icons/send_arrow.svg";
 
-import getTakerDetail from "../../helpMeDetailLayout/apis/getTakerDetail";
 import getMyInfo from "../../myPageEditLayout/apis/getMyInfo";
 import postHelpMeRegister from "../apis/postHelpMeRegister";
-import patchHelpMeRegister from "../apis/patchHelpMeRegister";
 import { helpMeFormData } from "../types";
 
 const cn = classNames.bind(styles);
@@ -58,12 +56,9 @@ const registerSchema = z.object({
 
 export default function HelpMeRegisterLayout() {
   const router = useRouter();
-  const queryClient = useQueryClient();
-  const { query } = router;
-  const isEditMode = Boolean(query.id);
   const isMountedRef = useRef(false);
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-  const [content, setContent] = useState<helpMeFormData | null | Partial<helpMeFormData>>(null);
+  const [content, setContent] = useState<helpMeFormData | null>(null);
 
   const { data: myInfoData, isFetching } = useQuery({
     queryKey: ["userInfo"],
@@ -78,91 +73,34 @@ export default function HelpMeRegisterLayout() {
     formState: { errors, isValid },
   } = useForm<helpMeFormData>({ resolver: zodResolver(registerSchema), mode: "onChange" });
 
-  const { data: prevData, isPending } = useQuery({
-    queryKey: ["takerDetail", query.id],
-    queryFn: () => getTakerDetail(query.id as string),
-    enabled: isEditMode,
-  });
-
   const uploadHelpMeMutation = useMutation({
     mutationFn: (content: helpMeFormData) => postHelpMeRegister(content),
     onSuccess: () => {
-      router.push(ROUTE.HELP_ME, undefined, { shallow:true });
+      router.push(ROUTE.HELP_ME);
     },
   });
 
-  const updateHelpMeMutation = useMutation({
-    mutationFn: (content: helpMeFormData) => patchHelpMeRegister(content, query.id as string),
-    onError: () =>{
-      console.log(errors);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["takerDetail", query.id] });
-      router.push(ROUTE.HELP_ME + "/" + query.id as string);
-    },
-  });
-
-  useEffect(() => {
-    if (prevData && !isPending) {
-      const prevHelpMeData: helpMeFormData = {
-        title: prevData.post.title,
-        assistanceType: prevData.post.assistance.assistanceType,
-        startDate: new Date(prevData.post.schedule.startDate),
-        endDate: new Date(prevData.post.schedule.endDate),
-        scheduleType: prevData.post.schedule.scheduleType,
-        scheduleDetails: prevData.post.schedule.scheduleDetails,
-        district: prevData.post.district,
-        content: prevData.post.content,
-        postType: prevData.postType,
-        gender: prevData.author.gender,
-        age: prevData.author.age,
-        disabilityType: prevData.author.disabilityType,
-        assistanceStartTime: prevData.post.assistance.assistanceStartTime,
-        assistanceEndTime: prevData.post.assistance.assistanceEndTime,
-      };
-
-      Object.entries(prevHelpMeData).forEach(([key, value]) => {
-        setValue(key as keyof helpMeFormData, value as never);
-      });
-    }
-  }, [prevData, isPending, setValue]);
-
-  const handleHelpMeSubmit = (data: helpMeFormData) => {
+  const handleHelpMetUpload = (data: helpMeFormData) => {
     setIsModalOpen((prev) => !prev);
 
     const content = {
-      ...data,
+      title: data.title,
+      assistanceType: data.assistanceType,
+      startDate: data.startDate,
+      endDate: data.endDate,
+      scheduleType: data.scheduleType,
+      scheduleDetails: data.scheduleDetails,
+      district: data.district,
+      content: data.content,
       postType: "TAKER",
       gender: myInfoData.gender,
       age: Number(myInfoData.age),
       disabilityType: myInfoData.disabilityType,
+      assistanceStartTime: data.assistanceStartTime,
+      assistanceEndTime: data.assistanceEndTime,
     };
 
-    if (isEditMode) {
-      const modifiedContent: Partial<helpMeFormData> = {};
-  
-      Object.entries(data).forEach(([key, value]) => {
-        const typedKey = key as keyof helpMeFormData;
-
-        if (value instanceof Date) {
-          const prevDate = new Date(prevData.post.schedule[typedKey]);
-          if (value.getTime() !== prevDate.getTime()) {
-            modifiedContent[typedKey] = value;
-          }
-        } else {
-          if (
-            value !== prevData.post[typedKey] &&
-            value !== prevData.post.assistance[typedKey] &&
-            value !== prevData.post.schedule[typedKey]
-          ) {
-            modifiedContent[typedKey] = value;
-          }
-        }
-      });
-      setContent(modifiedContent);
-    } else {
-      setContent(content);
-    }
+    setContent(content);
   };
 
   useEffect(() => {
@@ -187,7 +125,7 @@ export default function HelpMeRegisterLayout() {
         <div className={cn("box")}>
           <p className={cn("title")}>도와줄래요? 게시글 작성</p>
           <MyInfoCard />
-          <form className={cn("form")} onSubmit={handleSubmit(handleHelpMeSubmit)}>
+          <form className={cn("form")} onSubmit={handleSubmit(handleHelpMetUpload)}>
             <div className={cn("formContentBox")}>
               <div className={cn("titleContainer")}>
                 <Label className={cn("label")} htmlFor="title">
@@ -343,7 +281,7 @@ ex, 2시에 전대치과병원에서 진료 이동 도움이 필요합니다."
                 {errors.content && <p className={cn("errorMessage")}>{errors.content.message}</p>}
               </div>
               <Button className={cn("registerBox", { active: isValid })}>
-                {isEditMode ? "수정하기" : "등록하기"}
+                등록하기
                 <RegisterArrow className={cn("arrow")} />
               </Button>
             </div>
@@ -354,7 +292,7 @@ ex, 2시에 전대치과병원에서 진료 이동 도움이 필요합니다."
         <ConfirmModal
           setState={setIsModalOpen}
           content={content as helpMeFormData}
-          mutate={isEditMode ? updateHelpMeMutation.mutate : uploadHelpMeMutation.mutate}
+          mutate={uploadHelpMeMutation.mutate}
         />
       )}
     </>
