+++ src/pages/help-me/edit/index.tsx
@@ -1,12 +1,12 @@
 import { ReactElement } from "react";
 
 import RootLayout from "@/components/common/RootLayout/RootLayout";
-import HelpMeRegisterLayout from "@/components/page-layout/helpMeRegisterLayout/components/helpMeRegisterLayout";
+import HelpMeEditLayout from "@/components/page-layout/helpMeEditLayout/components/helpMeEdit";
 
 HelpMeEdit.getLayout = function getLayout(page: ReactElement) {
   return <RootLayout>{page}</RootLayout>;
 };
 
 export default function HelpMeEdit() {
-  return <HelpMeRegisterLayout />;
+  return <HelpMeEditLayout />;
 }
