+++ src/pages/help-you/edit/index.tsx
@@ -1,12 +1,12 @@
 import { ReactElement } from "react";
 
 import RootLayout from "@/components/common/RootLayout/RootLayout";
-import HelpYouRegisterLayout from "@/components/page-layout/helpYouRegisterLayout/components/helpYouRegisterLayout";
+import HelpYouEditLayout from "@/components/page-layout/helpYouEditLayout/components/helpYouEdit";
 
 HelpYouEdit.getLayout = function getLayout(page: ReactElement) {
   return <RootLayout>{page}</RootLayout>;
 };
 
 export default function HelpYouEdit() {
-  return <HelpYouRegisterLayout />;
+  return <HelpYouEditLayout />;
 }
